feat: show Popup when clicking icon
