+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require("path");
 
 module.exports = {
   entry: {
-    popup: path.join(__dirname, "src/popup.ts"),
+    popup: path.join(__dirname, "src/popup/index.tsx"),
     options: path.join(__dirname, "src/options.ts")
   },
   devtool: "inline-source-map",
